+++ backend/gdhub/Dtos/GameDetailsDto.cs
@@ -5,7 +5,7 @@ public record GameDetailsDto
     int Id,
     string Name,
     string Owner,
-    int Rating,
+    float Rating,
     string PicUrl,
     string Details
 );
